break setupConnection cycle
